Rewrote the fire plugin so that it scales to different strip lengths. Parameters dissipation and combustion are now 0-255 with suggested values at 128. Also fixed the EEPROM bootstrapping so that it works with the Core and uninitialized devices. Corrected a bug in the juggle plugin that made it fail whenever the micros timer wrapped around.

+++ Particle/plugin.h
@@ -24,10 +24,10 @@ class Plugin
 class FirePlugin : public Plugin
 {
     private:
-        uint32_t _dissipation;
-        uint32_t _combustion;
+        float _dissipation;
+        float _combustion;
 
-        uint8_t * _flames;
+        float * _flames;
 
         uint32_t _palette[256];
